cleanup pydantic
